ferrolearn-preprocess: StandardScaler float64 stats accumulators + per-op f32 transform rounding (closes #2205 #2305)
UPSTREAM SCIKIT-LEARN FILES OPENED:
  sklearn/preprocessing/_data.py:1016 (StandardScaler fit), :1064-1067 (transform
  X-=mean_ then X/=scale_ in-place f32), :1106-1109 (inverse X*=scale_ then
  X+=mean_), sklearn/utils/extmath.py _incremental_mean_and_var (float64
  accumulation, returns float64 mean_/var_ even for float32 input).

DESIGN DOC READ: .design/preprocess/standard_scaler.md

REQ STATUS (preprocess StandardScaler f32 numerical parity — R-DEV-1/R-DEV-3):
  StandardScaler f32 mean_/var_/scale_ float64 accumulators  divergent -> FIXED (#2205)
  StandardScaler<f32> transform per-op f32 rounding  divergent -> FIXED (#2305)

  acto-critic re-audit of the preprocess scalers' (already-SHIPPED) NaN handling
  surfaced two f32 divergences:
  - #2205: StandardScaler computed mean_/var_/scale_ in the generic F; sklearn ALWAYS
    accumulates in float64 (and stores float64 attributes) even for float32 input.
    Fixed: FittedStandardScaler stores mean_/var_/scale_ as Array1<f64> (matching
    sklearn's float64 attribute dtype); fit accumulates the NaN-skipping finite sum +
    squared-dev in f64; transform/inverse apply the f64 stats then downcast to F.
    The NaN-skip (REQ-7), finite-count divisor, all-NaN->NaN, inf-rejection, and
    zero-variance scale_=1 are preserved. f64 path bit-identical (50-fixture verify).
  - #2305: StandardScaler<f32>::transform fused (x-mean)/scale with ONE f32 rounding;
    sklearn does X-=mean_ then X/=scale_ in-place -> TWO f32 roundings (1 ULP near
    2^24). Fixed: two-step per-op rounding in transform (sub-then-round, div-then-
    round) + inverse_transform (REVERSE order mul-then-round, add-then-round), with
    the with_mean/with_std conditional skips + NaN passthrough preserved. f64
    identical (F::from(f64) exact).

VERIFICATION:
  cargo test -p ferrolearn-preprocess: 0 failed (divergence_standard_scaler 23 +
    _f32_transform_rounding 1 — f32 mean_ float64-exact, f32 transform exact bits
    [-0.7071068286895752,...,1.4142136573791504], f32 inverse round-trip, f64
    bit-identical, NaN/inf/zero-variance preserved); ferrolearn-python + meta-crate
    compile (getters now &Array1<f64>, PyO3 already f64)
  cargo clippy -p ferrolearn-preprocess --all-targets -- -D warnings: clean
  cargo fmt --all --check: clean
  Live sklearn 1.5.2 parity. RobustScaler shares the f32 class (#2206, follow-up);
  MaxAbsScaler does not (max-abs only, no accumulation).

Reference: scikit-learn 1.5.2 (commit 156ef14)

+++ ferrolearn-preprocess/tests/divergence_standard_scaler.rs
@@ -611,7 +611,8 @@ fn req7_f32_constant_column_scale_one() -> Result<(), FerroError> {
     let scaler = StandardScaler::<f32>::new();
     let x = array![[5.0f32, 1.0], [5.0, 2.0], [5.0, 3.0]];
     let fitted = scaler.fit(&x, &())?;
-    assert_eq!(fitted.scale()[0], 1.0f32);
+    // `scale_` is float64 even for f32 input (sklearn float64 attributes, #2205).
+    assert_eq!(fitted.scale()[0], 1.0f64);
     let scaled = fitted.transform(&x)?;
     for i in 0..3 {
         assert!(scaled[[i, 0]].abs() < 1e-6);
@@ -654,7 +655,6 @@ fn req7_f32_constant_column_scale_one() -> Result<(), FerroError> {
 ///
 /// Tracking: #2205
 #[test]
-#[ignore = "divergence: StandardScaler<f32> computes mean_/var_/scale_ in f32 not float64 accumulators; tracking #2205"]
 #[allow(
     clippy::approx_constant,
     reason = "these are the LIVE sklearn 1.5.2 f32 transform-output oracle values \
@@ -676,8 +676,9 @@ fn divergence_f32_uses_float64_accumulators() {
     let fitted = scaler.fit(&x, &()).unwrap();
 
     // sklearn's mean_ (computed in float64) is 16777217.333..., NOT the f32
-    // value 16777218.0 that ferrolearn computes.
-    let ferro_mean = f64::from(fitted.mean()[0]);
+    // value 16777218.0 that ferrolearn previously computed. `mean()` now returns
+    // f64 (sklearn's float64 attribute, #2205).
+    let ferro_mean = fitted.mean()[0];
     assert!(
         (ferro_mean - SK_MEAN).abs() < 0.5,
         "mean_: ferrolearn (f32) ={ferro_mean}, sklearn (float64) ={SK_MEAN}"
